Atualiza configuração do ambiente para RabbitMQ; refatora entidades e conversores para utilizar tipos curtos; implementa notificação de transferências via RabbitMQ; ajusta propriedades de conexão e adiciona novos arquivos de payload e exceção.

+++ core/src/main/java/com/eliasmeyer/sp/core/domain/model/carteira/TipoConta.java
@@ -2,16 +2,16 @@ package com.eliasmeyer.sp.core.domain.model.carteira;
 
 public enum TipoConta {
 
-	COMUM(1),
-	LOJISTA(2);
+	COMUM((short) 1),
+	LOJISTA((short) 2);
 
-	private final int codigo;
+	private final Short codigo;
 
-	TipoConta(int codigo) {
+	TipoConta(Short codigo) {
 		this.codigo = codigo;
 	}
 
-	public int getCodigo() {
+	public Short getCodigo() {
 		return codigo;
 	}
 }
